moved our tests to fix circular import and fixed silly little failing test
