+++ app/logic/participants.py
@@ -161,3 +161,33 @@ def getParticipationStatusForTrainings(program, userList, term):
             userParticipationStatus[user.username] = userParticipationStatus.get(user.username, []) + [(training, user.username in attendeeList)]
 
     return userParticipationStatus
+
+def sortParticipantsByStatus(event):
+    """
+    Takes in an event object, queries all participants, and then filters those
+    participants by their attendee status.
+
+    return: a list of participants who didn't attend, a list of participants who are waitlisted,
+    a list of participants who attended, and a list of all participants who have some status for the 
+    event.
+    """
+    eventParticipants = getEventParticipants(event)
+
+    # get all RSVPs for event and filter out those that did not attend into separate list
+    eventRsvpData = list(EventRsvp.select(EventRsvp, User).join(User).where(EventRsvp.event==event).order_by(EventRsvp.rsvpTime))
+    eventNonAttendedData = [rsvp for rsvp in eventRsvpData if rsvp.user not in eventParticipants]
+
+    if event.isPast:
+        eventVolunteerData = eventParticipants
+        
+        # if the event date has passed disregard the waitlist
+        eventWaitlistData = []
+    else:
+        # if rsvp is required for the event, grab all volunteers that are in the waitlist
+        eventWaitlistData = [volunteer for volunteer in (eventParticipants + eventRsvpData) if volunteer.rsvpWaitlist and event.isRsvpRequired]
+        
+        # put the rest of the users that are not on the waitlist into the volunteer data
+        eventVolunteerData = [volunteer for volunteer in eventNonAttendedData if volunteer not in eventWaitlistData]
+        eventNonAttendedData = []
+    
+    return eventNonAttendedData, eventWaitlistData, eventVolunteerData, eventParticipants
