+++ app/logic/volunteers.py
@@ -5,7 +5,6 @@ from app.models.eventRsvp import EventRsvp
 from app.models.program import Program
 from app.models.backgroundCheck import BackgroundCheck
 from app.models.programManager import ProgramManager
-from app.logic.participants import getEventParticipants
 from datetime import datetime, date
 from app.logic.createLogs import createAdminLog
 
@@ -106,33 +105,3 @@ def setProgramManager(username, program_id, action):
         studentstaff.addProgramManager(program_id)
     elif action == "remove":
         studentstaff.removeProgramManager(program_id)
-
-def sortParticipantsByStatus(event):
-    """
-    Takes in an event object, queries all participants, and then filters those
-    participants by their attendee status.
-
-    return: a list of participants who didn't attend, a list of participants who are waitlisted,
-    a list of participants who attended, and a list of all participants who have some status for the 
-    event.
-    """
-    eventParticipants = getEventParticipants(event)
-
-    # get all RSVPs for event and filter out those that did not attend into separate list
-    eventRsvpData = list(EventRsvp.select(EventRsvp, User).join(User).where(EventRsvp.event==event).order_by(EventRsvp.rsvpTime))
-    eventNonAttendedData = [rsvp for rsvp in eventRsvpData if rsvp.user not in eventParticipants]
-
-    if event.isPast:
-        eventVolunteerData = eventParticipants
-        
-        # if the event date has passed disregard the waitlist
-        eventWaitlistData = []
-    else:
-        # if rsvp is required for the event, grab all volunteers that are in the waitlist
-        eventWaitlistData = [volunteer for volunteer in (eventParticipants + eventRsvpData) if volunteer.rsvpWaitlist and event.isRsvpRequired]
-        
-        # put the rest of the users that are not on the waitlist into the volunteer data
-        eventVolunteerData = [volunteer for volunteer in eventNonAttendedData if volunteer not in eventWaitlistData]
-        eventNonAttendedData = []
-    
-    return eventNonAttendedData, eventWaitlistData, eventVolunteerData, eventParticipants
